Generalize the controlled access mixin
As it is also used with global preferences, thus without a tournament.

The tournament argument for `is_page_enabled` is also un-needed as the
object's tournament property is available there.
